+++ CHANGELOG.md
@@ -1,6 +1,9 @@
-PR - 10/13/2021
+PR Add form data functionality - 10/14/2021
 
 - Add api route and functionality to save form
+- Add api route and functionality to fetch forms
+- Add form display to user profile
+- Fix misc bugs
 
 PR Add user profile, logout - 10/13/2021